add simple two tailed t tests
here we are testing (on a small number of test runs)
which eBPF traces are significantly different in terms
of time spent. I need to run this again with at least 32
samples for each, and then I want to try a sleep wrapper
that allows me to capture the PID, and then a more advanced
approach with a LD preload thing.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>