ebpf: add automation to determine functions of interest
This includes a lima vm, along with a simple
set of steps to generate all the functions and run
the script against it (and time the whole thing).
This is not perfect but it is simple enough to
understand and use, and we can use it for our
further experiments!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>